requirements.txt
langchain
langchain-community
langchain-classic
langchain-core
langchain-groq
langchain-huggingface
langchain-text-splitters
duckduckgo-search
python-dotenv
pinecone-client
langchain-pinecone
sentence-transformers
pypdf
streamlit